Suggested modifications and comments left in relation to CM-1249 that found Tuxedo services running volumes without encryption.
----------------------------------------------------

encryption enabled on both the root volume and dynamic ebs volumes.

looks to be widespread use of the aws/ebs default KMS key so omitted the kms_key_id field.

With all the Tuxedo AMIs being single volume, the dynamic block doesn't look necessary so comment left to that effect.